back/requirements.txt
fastapi==0.111.1
pydantic==2.8.2
psycopg2==2.9.9
sqlalchemy==2.0.30
python-dotenv==1.0.1